#281 rename: Card -> CardTable folder
